feat: 중복된 fixture 삭제 및 공개 캡슐 조회 테스트 수정
