+++ backend/core/src/test/java/site/timecapsulearchive/core/common/fixture/VideoFixture.java
@@ -1,18 +1,18 @@
-package site.timecapsulearchive.core.common.data.fixture;
+package site.timecapsulearchive.core.common.fixture;
 
 import java.util.List;
 import java.util.stream.IntStream;
 import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
-import site.timecapsulearchive.core.domain.capsule.entity.Image;
+import site.timecapsulearchive.core.domain.capsule.entity.Video;
 import site.timecapsulearchive.core.domain.member.entity.Member;
 
-public class ImageFixture {
+public class VideoFixture {
 
-    public static List<Image> testImages(Member member, Capsule capsule, int count) {
+    public static List<Video> testVideos(Member member, Capsule capsule, int count) {
         return IntStream.range(0, count)
             .mapToObj(
-                dataPrefix -> Image.builder()
-                    .imageUrl(dataPrefix + "testImageURl")
+                dataPrefix -> Video.builder()
+                    .videoUrl(dataPrefix + "testVideoURl")
                     .member(member)
                     .capsule(capsule)
                     .build()
