+++ backend/core/src/test/java/site/timecapsulearchive/core/domain/capsule/media/repository/MediaQueryRepositoryTest.java
@@ -15,12 +15,13 @@ import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.TestConstructor.AutowireMode;
 import org.springframework.transaction.annotation.Transactional;
 import site.timecapsulearchive.core.common.RepositoryTest;
-import site.timecapsulearchive.core.common.data.fixture.CapsuleFixture;
-import site.timecapsulearchive.core.common.data.fixture.CapsuleSkinFixture;
-import site.timecapsulearchive.core.common.data.fixture.ImageFixture;
-import site.timecapsulearchive.core.common.data.fixture.MemberFixture;
-import site.timecapsulearchive.core.common.data.fixture.VideoFixture;
+import site.timecapsulearchive.core.common.fixture.CapsuleFixture;
+import site.timecapsulearchive.core.common.fixture.CapsuleSkinFixture;
+import site.timecapsulearchive.core.common.fixture.ImageFixture;
+import site.timecapsulearchive.core.common.fixture.MemberFixture;
+import site.timecapsulearchive.core.common.fixture.VideoFixture;
 import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
+import site.timecapsulearchive.core.domain.capsule.entity.CapsuleType;
 import site.timecapsulearchive.core.domain.capsule.entity.Image;
 import site.timecapsulearchive.core.domain.capsule.entity.Video;
 import site.timecapsulearchive.core.domain.capsule.repository.ImageQueryRepository;
@@ -52,13 +53,13 @@ class MediaQueryRepositoryTest extends RepositoryTest {
     @Transactional
     @BeforeEach
     void setUp(@Autowired EntityManager entityManager) {
-        member = MemberFixture.testMember(1);
+        member = MemberFixture.member(1);
         entityManager.persist(member);
 
-        CapsuleSkin capsuleSkin = CapsuleSkinFixture.testCapsuleSkin(1, member);
+        CapsuleSkin capsuleSkin = CapsuleSkinFixture.capsuleSkin(member);
         entityManager.persist(capsuleSkin);
 
-        capsule = CapsuleFixture.testCapsule(1, member, capsuleSkin);
+        capsule = CapsuleFixture.capsule(member, capsuleSkin, CapsuleType.PUBLIC);
         entityManager.persist(capsule);
 
         images = ImageFixture.testImages(member, capsule, 5);
